Add validations and secure password check
Previously there was a users table, but no User model to populate
the table with instances/objects. Therefore a User model was created
with presence validations for email, first name, last name, password,
and zipcode. There is also a uniqueness validation for an email
address so that there aren't multiple accounts with the same email.