Add method to downcase emails before saving
Previously user emails would be saved in their original inputted
format, which would mean that they would forever be case-sensitive,
which we don't necessarily want. So a method was created to downcase
all emails before saving them so that they are no longer case-sensitive
and cause less problems for logging in.