Update User create action
The create action for the User model was previously empty and not
performing any tasks. So the action was updated to now create a new
instance of a User with the permitted user params. If the instance
is valid and saved then the page returns a json of the user's
information. If the user is invalid then the page returns a JSON
with the current errors and correct status for wrong input