Define user_params method
The create action for the User controller was calling a user_params
method that had not yet been defined. So this method was defined and
it tells the program which params are permitted to be entered by the
user so that breakout hacking code cannot be entered into the fields
and corrupt the website/application. The method only accepts the
parameters defined as attributes for the User model.