Add method to set current user
Many of actions in the Users controller needs to know who the current
user that is logged in is, but it is repetitive to include this same
couple lines of code inside each action. So a private set_user method
was defined that sets the instance variable for a user to the user
whose id is included in the current session. This method is then
called at the top of the model to be performed before the actions
show, update, and destroy