Create Sessions Controller
The general way to keep track of the current user inside of a Rails
application is to use sessions which can store a piece of information,
typically the id, about the user that can be accessed at any point to
find the current user. Therefore the a Sessions Controller is used to
create new sessions for logging in, and destroy sessions for logging
out.