Define Sessions create action
The application will be using sessions to keep track of which user
is logged in, but there were no actions defined for the Sessions
Controller to do so. Therefore the create action was defined for
Sessions, which finds a user find the email inside of params and
then checks the password entered against the password saved in the
database for that user. If the passwords match then the id saved
within the session is set to that of the user. Else, JSON is rendered
containing an error saying that the email or password is invalid.