Define Sessions #destroy
Previously there was a create action for Sessions which would log
a user in, but there was no destroy action defined to log a user
out. So the action was defined with a simple line to reset the
session which will delete all information associated with the
current session and create a new blank one.