Fix custom route syntax
The custom routes that were written for the sessions controller
were not done correctly and caused errors when trying to run the
rails console. These errors were due to '/' missing from the beginning
of each route and a ':' missing after each to. Therefore both of
these things were added to the two custom route configurations and
the errors no longer arise.