Create migration for Goals-Users
There was already a join table for goals and users, but it was not
setup properly. Rails automatically assumes that you will follow
the convention of putting the two items for the join table in
alphabetical order. However, the join table already created was
UsersGoals, which the Rails app would not look for when trying to
associate users and goal, it would be looking for GoalsUser and it
did not exist, thus giving an error for a missing table. So a migration
containg the same info as the UsersGoals table was added and migrated,
but the table was titled GoalsUsers